i386: fix compilation warning
Some gcc versions complain about assigning long* to uint32_t* on i386.

* syscall.c [I386] (i386_esp_ptr): Change type from uint32_t* to long*.
* linux/i386/arch_regs.h (i386_esp_ptr): Likewise.
* linux/x86_64/arch_regs.h: Do not include "i386/arch_regs.h".
(i386_esp_ptr): New prototype.

+++ linux/i386/arch_regs.h
@@ -1,1 +1,1 @@
-extern uint32_t *const i386_esp_ptr;
+extern long *const i386_esp_ptr;
